+++ packages/plugin/webpack/src/WebpackPlugin.ts
@@ -99,23 +99,23 @@ export default class WebpackPlugin extends PluginBase<WebpackPluginConfig> {
     return null;
   }
 
-  getMainConfig = async () => {
-    const mainConfig = this.resolveConfig(this.config.mainConfig);
-
-    if (!mainConfig.entry) {
-      throw new Error('Required config option "entry" has not been defined');
-    }
-
+  getDefines = (upOneMore = false) => {
     const defines: { [key: string]: string; } = {};
-    let index = 0;
     if (!this.config.renderer.entryPoints || !Array.isArray(this.config.renderer.entryPoints)) {
       throw new Error('Required config option "renderer.entryPoints" has not been defined');
     }
     for (const entryPoint of this.config.renderer.entryPoints) {
-      defines[`${entryPoint.name.toUpperCase().replace(/ /g, '_')}_WEBPACK_ENTRY`] =
-        this.isProd
-        ? `\`file://\$\{require('path').resolve(__dirname, '../renderer', '${entryPoint.name}', 'index.html')\}\``
-        : `'http://localhost:${BASE_PORT + index}'`;
+      if (entryPoint.html) {
+        defines[`${entryPoint.name.toUpperCase().replace(/ /g, '_')}_WEBPACK_ENTRY`] =
+          this.isProd
+          ? `\`file://\$\{require('path').resolve(__dirname, '../renderer', '${upOneMore ? '..' : '.'}', '${entryPoint.name}', 'index.html')\}\``
+          : `'http://localhost:${BASE_PORT}/${entryPoint.name}'`;
+      } else {
+        defines[`${entryPoint.name.toUpperCase().replace(/ /g, '_')}_WEBPACK_ENTRY`] =
+          this.isProd
+          ? `\`file://\$\{require('path').resolve(__dirname, '../renderer', '${upOneMore ? '..' : '.'}', '${entryPoint.name}', 'index.js')\}\``
+          : `'http://localhost:${BASE_PORT}/${entryPoint.name}/index.js'`;
+      }
 
       if (entryPoint.preload) {
         defines[`${entryPoint.name.toUpperCase().replace(/ /g, '_')}_PRELOAD_WEBPACK_ENTRY`] =
@@ -123,11 +123,22 @@ export default class WebpackPlugin extends PluginBase<WebpackPluginConfig> {
           ? `\`file://\$\{require('path').resolve(__dirname, '../renderer', '${entryPoint.name}', 'preload.js')\}\``
           : `'${path.resolve(this.baseDir, 'renderer', entryPoint.name, 'preload.js')}'`;
       }
-      index += 1;
     }
+    return defines;
+  }
+
+  getMainConfig = async () => {
+    const mainConfig = this.resolveConfig(this.config.mainConfig);
+
+    if (!mainConfig.entry) {
+      throw new Error('Required config option "entry" has not been defined');
+    }
+
+    const defines = this.getDefines();
     return merge.smart({
       devtool: 'source-map',
       target: 'electron-main',
+      mode: this.isProd ? 'production' : 'development',
       output: {
         path: path.resolve(this.baseDir, 'main'),
         filename: 'index.js',
@@ -157,6 +168,7 @@ export default class WebpackPlugin extends PluginBase<WebpackPluginConfig> {
     return merge.smart({
       devtool: 'inline-source-map',
       target: 'electron-renderer',
+      mode: this.isProd ? 'production' : 'development',
       entry: prefixedEntries.concat([
         entryPoint.js,
       ]),
@@ -171,29 +183,39 @@ export default class WebpackPlugin extends PluginBase<WebpackPluginConfig> {
     }, rendererConfig);
   }
 
-  getRendererConfig = async (entryPoint: WebpackPluginEntryPoint) => {
+  getRendererConfig = async (entryPoints: WebpackPluginEntryPoint[]) => {
     const rendererConfig = this.resolveConfig(this.config.renderer.config);
-    const prefixedEntries = entryPoint.prefixedEntries || [];
+    const entry: webpack.Entry = {};
+    for (const entryPoint of entryPoints) {
+      const prefixedEntries = entryPoint.prefixedEntries || [];
+      entry[entryPoint.name] = prefixedEntries
+        .concat([entryPoint.js])
+        .concat(this.isProd || !Boolean(entryPoint.html) ? [] : ['webpack-hot-middleware/client']);
+    }
+
+    const defines = this.getDefines(true);
     return merge.smart({
+      entry,
       devtool: 'inline-source-map',
       target: 'electron-renderer',
-      entry: prefixedEntries.concat([
-        entryPoint.js,
-      ]).concat(this.isProd ? [] : ['webpack-hot-middleware/client']),
+      mode: this.isProd ? 'production' : 'development',
       output: {
-        path: path.resolve(this.baseDir, 'renderer', entryPoint.name),
-        filename: 'index.js',
+        path: path.resolve(this.baseDir, 'renderer'),
+        filename: '[name]/index.js',
+        globalObject: 'self',
       },
       node: {
         __dirname: false,
         __filename: false,
       },
-      plugins: [
+      plugins: entryPoints.filter(entryPoint => Boolean(entryPoint.html)).map(entryPoint =>
         new HtmlWebpackPlugin({
           title: entryPoint.name,
           template: entryPoint.html,
+          filename: `${entryPoint.name}/index.html`,
+          chunks: [entryPoint.name].concat(entryPoint.additionalChunks || []),
         }),
-      ].concat(this.isProd ? [] : [new webpack.HotModuleReplacementPlugin()]),
+      ).concat([new webpack.DefinePlugin(defines)]).concat(this.isProd ? [] : [new webpack.HotModuleReplacementPlugin()]),
     }, rendererConfig);
   }
 
@@ -226,54 +248,53 @@ export default class WebpackPlugin extends PluginBase<WebpackPluginConfig> {
   }
 
   compileRenderers = async (watch = false) => {
-    for (const entryPoint of this.config.renderer.entryPoints) {
-      await asyncOra(`Compiling Renderer Template: ${entryPoint.name}`, async () => {
-        await new Promise(async (resolve, reject) => {
-          webpack(await this.getRendererConfig(entryPoint)).run((err, stats) => {
-            if (err) return reject(err);
-            resolve();
-          });
+    await asyncOra('Compiling Renderer Template', async () => {
+      await new Promise(async (resolve, reject) => {
+        webpack(await this.getRendererConfig(this.config.renderer.entryPoints)).run((err, stats) => {
+          if (err) return reject(err);
+          resolve();
         });
-        if (entryPoint.preload) {
+      });
+    });
+
+    for (const entryPoint of this.config.renderer.entryPoints) {
+      if (entryPoint.preload) {
+        await asyncOra(`Compiling Renderer Preload: ${entryPoint.name}`, async () => {
           await new Promise(async (resolve, reject) => {
             webpack(await this.getPreloadRendererConfig(entryPoint, entryPoint.preload!)).run((err, stats) => {
               if (err) return reject(err);
               resolve();
             });
           });
-        }
-      });
+        });
+      }
     }
   }
 
   launchDevServers = async (logger: Logger) => {
     await asyncOra('Launch Dev Servers', async () => {
-      let index = 0;
-      for (const entryPoint of this.config.renderer.entryPoints) {
-        const tab = logger.createTab(entryPoint.name);
+      const tab = logger.createTab('Renderers');
 
-        const config = await this.getRendererConfig(entryPoint);
-        const compiler = webpack(config);
-        const server = webpackDevMiddleware(compiler, {
-          logger: {
-            log: tab.log.bind(tab),
-            info: tab.log.bind(tab),
-            error: tab.log.bind(tab),
-            warn: tab.log.bind(tab),
-          },
-          publicPath: '/',
-          hot: true,
-          historyApiFallback: true,
-        } as any);
-        const app = express();
-        app.use(server);
-        app.use(webpackHotMiddleware(compiler));
-        this.servers.push(app.listen(BASE_PORT + index));
-        index += 1;
-      }
+      const config = await this.getRendererConfig(this.config.renderer.entryPoints);
+      const compiler = webpack(config);
+      const server = webpackDevMiddleware(compiler, {
+        logger: {
+          log: tab.log.bind(tab),
+          info: tab.log.bind(tab),
+          error: tab.log.bind(tab),
+          warn: tab.log.bind(tab),
+        },
+        publicPath: '/',
+        hot: true,
+        historyApiFallback: true,
+      } as any);
+      const app = express();
+      app.use(server);
+      app.use(webpackHotMiddleware(compiler));
+      this.servers.push(app.listen(BASE_PORT));
     });
 
-    await asyncOra('Compile Preload Scripts', async () => {
+    await asyncOra('Compiling Preload Scripts', async () => {
       for (const entryPoint of this.config.renderer.entryPoints) {
         if (entryPoint.preload) {
           await new Promise(async (resolve, reject) => {